refactor(selection): name the filter files for what they do
qunit-filter.ts and qunit-filter-match.ts sat next to each other with
near-identical names for opposite roles, so a first reader could not tell which
was which:

- filter-query.ts  — builds the filter we SEND to the browser (isFilteredRun,
                     qunitFilterQuery, describeFilter)
- filter-match.ts  — the port of QUnit's matcher we EVALUATE locally, in
                     process, for --search (matchesQUnitFilter, qunitFullName)

Inside lib/selection/ the qunit- prefix was redundant; query-vs-match names the
actual distinction. Also socket-utils.ts -> socket-io.ts in the daemon: "utils"
said nothing about attachLineParser/probeSocket, which are socket plumbing.

File moves and import paths only — no behavior change.
