+++ lib/commands/daemon/socket-io.ts
