+++ test/selection/filter-match-test.ts
@@ -1,5 +1,5 @@
 import { module, test } from 'qunitx';
-import { matchesQUnitFilter, qunitFullName } from '../../lib/selection/qunit-filter-match.ts';
+import { matchesQUnitFilter, qunitFullName } from '../../lib/selection/filter-match.ts';
 
 // The exact fullNames QUnit builds for test/fixtures + the overlapping-name cart fixture. Every
 // expectation below was verified against a real browser run before being encoded here, so this
